Regularize package of jetty-hazelcast.
