+++ jetty-session/jetty-hazelcast/src/main/java/org/eclipse/jetty/session/hazelcast/SessionDataSerializer.java
@@ -11,7 +11,7 @@
 // ========================================================================
 //
 
-package org.eclipse.jetty.hazelcast.session;
+package org.eclipse.jetty.session.hazelcast;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
